pyproject.toml
[project]
name = "trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "groq>=0.37.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
    "requests>=2.31.0",
]
